[wip] Expose list of build assets under `/assets/-/:pkg/:env/:version?`
