+++ lib/preboots/index.js
@@ -22,6 +22,7 @@ module.exports = function (app, options, done) {
   app.preboot(require('./loading-bay'));
   app.preboot(require('./agents'));
   app.preboot(require('./carpenter'));
+  app.preboot(require('./spec'));
   app.preboot(require('./npm-auth'));
   app.preboot(require('../npm/preboot'));
