+++ lib/routes/assets.js
@@ -1,11 +1,11 @@
 /* eslint no-bitwise: 0*/
 'use strict';
 
-var comment = require('commenting'),
-  zipline = require('zipline'),
-  hyperquest = require('hyperquest'),
-  mime = require('mime'),
-  path = require('path');
+const comment = require('commenting');
+const zipline = require('zipline');
+const hyperquest = require('hyperquest');
+const mime = require('mime');
+const path = require('path');
 
 //
 // As our intention is to primary serve JS builds, this should be our default
@@ -14,6 +14,38 @@ var comment = require('commenting'),
 mime.default_type = 'text/javascript';
 
 module.exports = function (app) {
+  const spec = app.spec;
+
+  /*
+   * API for serving latest files
+   */
+  app.routes.get('/assets/-/:pkg/:env/:version?', function (req, res) {
+    res.setHeader('Content-Type', 'application/json');
+    res.statusCode = 200;
+
+    spec.findBuild(
+      spec.fromRequest(req),
+      function (err, build) {
+        if (err) { return next(err); }
+
+        //
+        // TODO (indexzero): proper sanitization on this untrusted
+        // user input.
+        //
+        const { filter } = req.query;
+        if (filter) {
+          res.send(
+            build.files.filter(file => {
+              return path.extname(file) === filter;
+            });
+          )
+        }
+
+        res.send(build.files);
+      }
+    );
+  });
+
   /**
    * API endpoint for serving builds.
    * TODO: Whats the right way to support gzip if our content is stored in an
