+++ lib/routes/builds.js
@@ -5,28 +5,9 @@ var url = require('url');
 var stringify = require('stringify-stream');
 
 const stringifyOpts = { open: '[', close: ']' };
-const defaultLocale = 'en-US';
-
-/**
- * Generate a build specification so we can fetch it from bffs.
- *
- * @param {Object} params Parameters.
- * @param {Object} query Query string args
- * @param {String} env Fallback environment in case it's missing.
- * @returns {Object} proper spec object
- * @api private
- */
-function spec(params, query, env) {
-  query = query || {};
-  return {
-    env: params.env || env,
-    version: params.version,
-    name: params.pkg,
-    locale: query.locale || defaultLocale
-  };
-}
 
 module.exports = function (app) {
+  const spec = app.spec;
   const auth = app.middlewares.auth;
   //
   // ### /builds/head
@@ -119,97 +100,15 @@ module.exports = function (app) {
   // environments if no env is specified.
   //
   app.routes.get('/builds/:pkg/:env/:version?', auth, function (req, res, next) {
-    var error;
-    var defaulted = false;
-
     res.setHeader('Content-Type', 'application/json');
     res.statusCode = 200;
 
-    //
-    // Normalzie the build object
-    //
-    function normalizeBuild(build) {
-
-      function normalizeUrl(filePath) {
-        return url.resolve(build.cdnUrl, filePath);
+    spec.findBuild(
+      spec.fromRequest(req),
+      function (err, build) {
+        if (err) { return next(err); }
+        res.send(build);
       }
-
-      build.files = build.recommended.length
-        ? build.recommended.map(normalizeUrl)
-        : build.artifacts.map(normalizeUrl);
-
-      return build;
-    }
-
-    //
-    // Ok so the assumption currently is that we will use the locale passed on
-    // the query string or we use en-US for all of the builds fetched.
-    //
-    var specs = spec(req.params, req.query);
-
-    return fetch(specs);
-
-    function notFound() {
-      error = new Error('Build not found');
-      error.status = 404;
-      return next(error);
-    }
-
-    function fetch(spec) {
-      //
-      // If we have a version, do search, otherwise fetch from HEAD,
-      // we may want to automatically handle this in `bffs`
-      //
-      return spec.version
-        ? app.bffs.search(spec, fetched)
-        : app.bffs.head(spec, fetched);
-    }
-
-    function fetched(err, build) {
-      if (err) {
-        err.status = 500;
-        return next(err);
-      }
-
-      //
-      // Remark: Do a secondary lookup when we have a locale if we get a miss the first
-      // time if we have a 2 part locale
-      //
-      if (!build && !specs.locale) return notFound();
-      else if (!build && specs.locale) {
-        //
-        // 1. Try and grab the iso369 of the locale and refetch
-        // 2. Otherwise try and fetch the defaultLocale
-        // 3. Return notFound if all else fails
-        const lang = iso639(specs.locale);
-        if (lang) return refetch(lang);
-        if (defaulted) return notFound();
-        defaulted = true;
-        return refetch(defaultLocale);
-      }
-
-      //
-      // Grab the relevant file info for the fingerprints on these builds
-      //
-      res.send(normalizeBuild(build.toJSON()));
-    }
-
-    //
-    // Retries fetch for build with given locale string
-    //
-    function refetch(locale) {
-      specs.locale = locale;
-      return fetch(specs);
-    }
+    )
   });
 };
-
-//
-// Returns back an iso639 string if given a full locale,
-// if given the same iso639 string, returns null
-//
-function iso639(locale) {
-  const parts = locale.split('-');
-  if (parts[0] === locale) return null;
-  return parts[0];
-}
